+++ 42cursus/0circle/libft/ft_memccpy.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_memccpy.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: seungnle <marvin@42.fr>                    +#+  +:+       +#+        */
+/*   By: seungnle <seungnle@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/10/08 03:02:46 by seungnle          #+#    #+#             */
-/*   Updated: 2020/10/08 03:02:48 by seungnle         ###   ########.fr       */
+/*   Updated: 2020/10/16 21:00:53 by seungnle         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@ void	*ft_memccpy(void *dst, const void *src, int c, size_t n)
 {
 	unsigned char		*a;
 	unsigned const char	*b;
-	int					i;
+	size_t				i;
 	unsigned char		chr;
 
 	chr = (unsigned char)c;
@@ -25,7 +25,7 @@ void	*ft_memccpy(void *dst, const void *src, int c, size_t n)
 		return (0);
 	a = dst;
 	b = src;
-	while (i < (int)n)
+	while (i < n)
 	{
 		a[i] = b[i];
 		if (b[i] == chr)
